+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changes
+* Load config into index.jsx statically instead of dynamically.
 
 0.4.1 - (April 11, 2018)
 ----------